[Issue-1050] [Confirmation] refactor import
