+++ packages/extension-koni-ui/src/Popup/Confirmations/Detail/Evm/index.tsx
@@ -1,0 +1,5 @@
+// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as EvmMessageDetail } from './Message';
+export { default as EvmTransactionDetail } from './Transaction';
