+++ packages/extension-koni-ui/src/Popup/Confirmations/Detail/index.tsx
@@ -1,0 +1,7 @@
+// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as BaseDetailModal } from './BaseDetailModal';
+
+export * from './Evm';
+export * from './Substrate';
