+++ packages/extension-koni-ui/src/Popup/Confirmations/EvmSignatureConfirmation.tsx
@@ -2,13 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { ConfirmationsQueueItem, EvmSignatureRequest } from '@subwallet/extension-base/background/KoniTypes';
-import AccountItemWithName from '@subwallet/extension-koni-ui/components/Account/Item/AccountItemWithName';
-import ConfirmationGeneralInfo from '@subwallet/extension-koni-ui/components/Confirmation/ConfirmationGeneralInfo';
-import ViewDetailIcon from '@subwallet/extension-koni-ui/components/Icon/ViewDetailIcon';
+import { AccountItemWithName, ConfirmationGeneralInfo, ViewDetailIcon } from '@subwallet/extension-koni-ui/components';
 import useOpenDetailModal from '@subwallet/extension-koni-ui/hooks/confirmation/useOpenDetailModal';
-import BaseDetailModal from '@subwallet/extension-koni-ui/Popup/Confirmations/Detail/BaseDetailModal';
-import EvmMessageDetail from '@subwallet/extension-koni-ui/Popup/Confirmations/Detail/Evm/Message';
-import EvmSignArea from '@subwallet/extension-koni-ui/Popup/Confirmations/Sign/Evm';
 import { ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { EvmSignatureSupportType } from '@subwallet/extension-koni-ui/types/confirmation';
 import { isEvmMessage } from '@subwallet/extension-koni-ui/util';
@@ -17,6 +12,9 @@ import CN from 'classnames';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
+
+import { BaseDetailModal, EvmMessageDetail } from './Detail';
+import { EvmSignArea } from './Sign';
 
 interface Props extends ThemeProps {
   type: EvmSignatureSupportType
