+++ packages/extension-koni-ui/src/Popup/Confirmations/Qr/index.tsx
@@ -1,0 +1,5 @@
+// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './DisplayPayload';
+export { default as ScanSignature } from './ScanSignature';
