+++ packages/extension-koni-ui/src/Popup/Confirmations/Sign/index.tsx
@@ -1,0 +1,5 @@
+// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as EvmSignArea } from './Evm';
+export { default as SubstrateSignArea } from './Substrate';
