+++ packages/extension-koni-ui/src/Popup/Confirmations/Transaction/index.tsx
@@ -5,11 +5,6 @@ import { ConfirmationDefinitions, ExtrinsicType } from '@subwallet/extension-bas
 import { SigningRequest } from '@subwallet/extension-base/background/types';
 import { SWTransactionResult } from '@subwallet/extension-base/services/transaction-service/types';
 import useParseSubstrateRequestPayload from '@subwallet/extension-koni-ui/hooks/confirmation/useParseSubstrateRequestPayload';
-import EvmSignArea from '@subwallet/extension-koni-ui/Popup/Confirmations/Sign/Evm';
-import SubstrateSignArea from '@subwallet/extension-koni-ui/Popup/Confirmations/Sign/Substrate';
-import BaseTransactionConfirmation from '@subwallet/extension-koni-ui/Popup/Confirmations/Transaction/parts/Base';
-import StakeTransactionConfirmation from '@subwallet/extension-koni-ui/Popup/Confirmations/Transaction/parts/Stake';
-import { TransferBlock } from '@subwallet/extension-koni-ui/Popup/Confirmations/Transaction/parts/TransferBlock';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { ConfirmationQueueItem } from '@subwallet/extension-koni-ui/stores/base/RequestState';
 import { ThemeProps } from '@subwallet/extension-koni-ui/types';
@@ -17,7 +12,9 @@ import CN from 'classnames';
 import React, { useCallback, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import SendNftTransactionConfirmation from "@subwallet/extension-koni-ui/Popup/Confirmations/Transaction/parts/SendNft";
+
+import { EvmSignArea, SubstrateSignArea } from '../Sign';
+import { BaseTransactionConfirmation, SendNftTransactionConfirmation, StakeTransactionConfirmation, TransferBlock } from './parts';
 
 interface Props extends ThemeProps {
   confirmation: ConfirmationQueueItem;
