+++ packages/extension-koni-ui/src/Popup/Confirmations/Transaction/parts/index.tsx
@@ -1,0 +1,9 @@
+// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export type { BaseTransactionConfirmationProps } from './Base';
+export { default as BaseTransactionConfirmation } from './Base';
+export { default as SendNftTransactionConfirmation } from './SendNft';
+export { default as StakeTransactionConfirmation } from './Stake';
+
+export * from './TransferBlock';
